requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
openpyxl
Pillow
